move useOrganizationsQuery to shared
