+++ src/Designer/frontend/packages/shared/src/hooks/queries/index.ts
@@ -16,3 +16,4 @@ export { useResourcePolicyActionsQuery } from './useResourcePolicyActionsQuery';
 export { useResourcePolicySubjectsQuery } from './useResourcePolicySubjectsQuery';
 export { useResourceAccessPackagesQuery } from './useResourceAccessPackagesQuery';
 export { useUserOrgPermissionsQuery } from './useUserOrgPermissionsQuery';
+export { useOrganizationsQuery } from './useOrganizationsQuery';
